Added information to clarify the Win11 requirement for TitleBar customization
Added comments in the code to make it clearer that the IsCustomizationSupported call is there due to the functionality only bein available on Windows 11 and later.
Added a note to the readme to call out the Windows 11 requirement for TitleBar customization and a link to the conceptual article for TitleBar customization as well.