There was a problem with Samples/Windowing/cs-winui when using the TitleTextBox.Text.Contains method to find a match.
Instead of displaying the expected error dialog, an exception was thrown with the error messages:

"Value does not fall within the expected range."
"The parameter is incorrect."
"This element is already associated with a XamlRoot, it cannot be associated with a different one until it is removed from the previous XamlRoot."

To fix this issue, the solution was to assign XamlRoot to this.XamlRoot. By making this assignment, the problem was resolved, and the exception was no longer thrown.

Since you cannot paste the "\n" into the text box as it is not a multiline input, this code is unreachable and not tested.